+++ client/src/components/flightplanslayer/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./FlightPlansLayer";
+export { SelectedFlightPlanLayer } from "./FlightPlansLayer";